- Issue #124, Cédric Luthi's fix for -[NSWindow _makeSureIsOnAScreen] - Improved path stroking in the GDI context, line cap, line join and dashes enabled, along with improved coordinate/size transformation - Fixed drawing cached image rep's in KGContext_builtin_gdi

+++ AppKit/CoreGraphics.subproj/KGGraphicsState.h
@@ -31,7 +31,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
    CGLineJoin          _lineJoin;
    float               _miterLimit;
    float               _dashPhase;
-   float               _dashLengthsCount;
+   int                 _dashLengthsCount;
    float              *_dashLengths;
    CGColorRenderingIntent _renderingIntent;
    CGBlendMode          _blendMode;
